api/requirements.txt
fastapi
langchain-openai
langchain
python-dotenv
uvicorn 

